:recycle: notice-스키마 수정
