+++ prisma/schema.prisma
@@ -25,6 +25,7 @@ model User {
   Token     Token[]
   Study     Study[]
   Schedule  Schedule[]
+  Notice    Notice[]
   createdAt DateTime   @default(now())
   updatedAt DateTime   @updatedAt
   // StudyTodo StudyTodo[]
@@ -39,7 +40,7 @@ model Schedule {
   UserId    Int
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
-}
+
 
 model Token {
   id        Int      @id @default(autoincrement())
@@ -61,11 +62,12 @@ model Study {
   membersLimit String?
   chatLink     String?
   image        String?
+  notice        Notice[]
   joinMember   Json?
   createdAt    DateTime    @default(now())
   updatedAt    DateTime    @updatedAt
   StudyTodo    StudyTodo[]
-}
+
 
 model Todo {
   id        Int       @id @default(autoincrement())
@@ -88,11 +90,14 @@ model StudyTodo {
 }
 
 model Notice {
-  id        Int      @id @default(autoincrement())
+  id        Int      @id @default(autoincrement()) 
   category  String
   title     String
   content   String
-  author    String?
+  author    User     @relation(fields: [authorId], references: [id], onDelete: Cascade)
+  authorId  Int 
+  study     Study    @relation(fields: [studyId], references: [id], onDelete: Cascade)
+  studyId   Int
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
 }
