+++ src/pages/api/notice/index.ts
@@ -1,15 +1,26 @@
 import { NextApiRequest, NextApiResponse } from 'next';
+import { StudyTimer } from 'src/components/StudyMain/StudyTimer';
 import client from 'src/libs/server/client';
 import withHandler from 'src/libs/server/withHandler';
+import { withApiSession } from 'src/libs/server/withSession';
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
   if (req.method === 'POST') {
-    const [{ category, title, content }] = req.body;
+    const { category, title, content } = req.body;
+    const { user } = req.session;
     let noticeData = await client.notice.create({
       data: {
         category: category,
         title: title,
         content: content,
+        author: {
+          connect: {
+            id: user?.id,
+          },
+        },
+        study: {
+          connect,
+        },
       },
     });
     return res.json({
@@ -20,6 +31,7 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
 
   if (req.method === 'GET') {
     const queryid = req.query.id;
+    console.log(queryid);
     if (queryid !== 'many') {
       let noticeData = await client.notice.findUnique({
         where: {
@@ -28,7 +40,11 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
       });
       return res.json({ ok: true, noticeData });
     } else {
-      let noticeData = await client.notice.findMany();
+      let noticeData = await client.notice.findMany({
+        include: {
+          study: true,
+        },
+      });
       return res.json({ ok: true, noticeData });
     }
   }
