+++ src/pages/api/notice/view.ts
@@ -1,6 +1,6 @@
 import useSWR from 'swr';
 
 export default function view(id: string | string[] | undefined) {
-  const { data } = useSWR(`/api//notice/?id=${id}`);
+  const { data } = useSWR(`/api/notice/?id=${id}`);
   return data;
 }
