+++ src/pages/study/[studyId]/notice/write.tsx
@@ -1,13 +1,12 @@
 import styled from '@emotion/styled';
-import { StudyBanner } from 'src/components/StudyMain/StudyBanner';
-import { COLOR } from 'src/constants';
+import { StudyBanner } from '../../../../components/StudyMain/StudyBanner';
+import { COLOR } from '../../../../constants';
 import dynamic from 'next/dynamic';
 import useMutation from 'src/libs/client/useMutation';
-import { NoticeTitle } from 'src/components/Notice/NoticeTitle';
-import { Button } from 'src/components/Notice/Button';
-import React, { useEffect, useState } from 'react';
-import { useRouter } from 'next/router';
-import view from '../../../api/notice/view';
+import { NoticeTitle } from '../../../../components/Notice/NoticeTitle';
+import { Button } from '../../../../components/Notice/Button';
+import { NoticeData } from 'src/types/Notice';
+import React, { useState } from 'react';
 
 const PostEditor = dynamic(
   () => import('../../../../components/Notice/PostEditor'),
@@ -16,30 +15,12 @@ const PostEditor = dynamic(
   },
 );
 
-interface NoticeData {
-  id: number;
-  category: string;
-  title: string;
-  content: string;
-}
-
 export default function NoticePage(): JSX.Element {
+  const [notice, { loading, data, error }] = useMutation('/api/notice');
   let [noticeList, setNoticeList] = useState<NoticeData[]>([]);
-  const router = useRouter();
-  const { id } = router.query;
-  const [beforeNotice, setBeforeNotice] = useState<NoticeData[]>([]);
-  const beforeNoticeData = view(id);
-  useEffect(() => {
-    setBeforeNotice(beforeNoticeData);
-  }, [beforeNoticeData]);
-
-  const [category, setCategory] = useState(
-    beforeNoticeData?.noticeData?.category,
-  );
-  const [title, setTitle] = useState(beforeNoticeData?.noticeData?.title);
-  const [content, setContent] = useState(beforeNoticeData?.noticeData?.content);
-
-  const [edit] = useMutation('/api/notice/edit');
+  const [category, setCategory] = useState('');
+  const [title, setTitle] = useState('');
+  const [content, setContent] = useState('');
 
   const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const {
@@ -57,14 +38,14 @@ export default function NoticePage(): JSX.Element {
     setContent(editor);
   };
 
-  const onSubmit = async (e: any) => {
+  const onSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
     setNoticeList(
-      (noticeList = [
-        { id: +id, category: category, title: title, content: content },
-      ]),
+      (noticeList = [{ category: category, title: title, content: content }]),
     );
     const data = noticeList;
-    edit(data);
+    notice(data);
+    reset();
   };
 
   const reset = () => {
@@ -86,7 +67,7 @@ export default function NoticePage(): JSX.Element {
         link="#"
       />
       <NoticeTitle />
-      <NoticeForm onSubmit={onSubmit} action="/study/notice/">
+      <NoticeForm onSubmit={onSubmit} action="/study/notice">
         <div className="list">
           <label htmlFor="input-category">말머리</label>
           <input
@@ -94,7 +75,7 @@ export default function NoticePage(): JSX.Element {
             list="category-list"
             id="input-category"
             name="category"
-            placeholder={`${beforeNoticeData?.noticeData?.category}`}
+            placeholder="최대 4자까지 입력할 수 있습니다."
           />
         </div>
         <div className="list">
@@ -105,7 +86,6 @@ export default function NoticePage(): JSX.Element {
             type="text"
             id="input-title"
             className="w100"
-            placeholder={`${beforeNoticeData?.noticeData?.title}`}
           />
         </div>
         <div className="list">
